pyproject.toml
[project]
name = "SIDE"
version = "0.1.0"
description = "Surgical Instrument Depth Estimation & Entity Segmentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "colorlog>=6.9.0",
    "einops>=0.8.2",
    "imageio[ffmpeg]>=2.37.2",
    "iprogress>=0.4",
    "ipykernel>=6.30.1",
    "mlflow==3.4.0",
    "monai>=1.5.1",
    "nbformat>=5.10.4",
    "numpy>=2.1.2",
    "omegaconf>=2.3.0",
    "open3d>=0.19.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.0.0",
    "plotly>=6.3.0",
    "pytest>=8.4.2",
    "pyvista>=0.46.5",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
    "timm>=1.0.24",
    "torch>=2.8.0",
    "torch-lr-finder>=0.2.2",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "trimesh>=4.11.2",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.pytest.ini_options]
pythonpath = ["."]
